level: merge Level_Initialises/Level_Unload in TRX
