+++ src/tr1/game/game/game.c
@@ -81,6 +81,7 @@ bool Game_Start(const GF_LEVEL *const level, const GF_SEQUENCE_CONTEXT seq_ctx)
 void Game_End(void)
 {
     Savegame_PersistGameToCurrentInfo(Game_GetCurrentLevel());
+    Music_Stop();
 }
 
 GF_COMMAND Game_Control(const bool demo_mode)
