--- src/tr1/game/objects/creatures/pierre.h
@@ -1,3 +1,0 @@
-#pragma once
-
-void Pierre_Reset(void);
