+++ src/tr1/global/vars.c
@@ -12,4 +12,4 @@ int32_t g_PhdBottom = 0;
 int32_t g_PhdRight = 0;
 int32_t g_PhdTop = 0;
 
-GAME_INFO g_GameInfo = { .select_save_slot = -1 };
+GAME_INFO g_GameInfo = { .select_save_slot = -1, .select_level_num = -1 };
