+++ src/tr2/game/demo.c
@@ -183,9 +183,7 @@ void Demo_End(void)
     M_PRIV *const p = &m_Priv;
     M_RestoreConfig(p);
     Overlay_SetBottomText(nullptr, false);
-    Sound_StopAll();
     Music_Stop();
-    Music_SetVolume(g_Config.audio.music_volume);
 }
 
 void Demo_Pause(void)
