+++ src/tr2/game/level.h
@@ -1,10 +1,6 @@
 #pragma once
 
-#include "game/game_flow/types.h"
-
 #include <libtrx/game/level.h>
 
 void Level_Init(void);
 bool Level_Initialise(const GF_LEVEL *level, GF_SEQUENCE_CONTEXT seq_ctx);
-void Level_Load(const GF_LEVEL *level);
-void Level_Unload(void);
